refactor: move PendingProcessingResult and ConversionStatus out of Services namespace
Interfaces must not depend on the implementation namespace. Both
records are now in Application/DTOs, and ICurrencyRateService only
references DTOs. Pure move, no runtime behavior change.
