+++ src/MyAccountingApp.Application/Interfaces/ICurrencyRateService.cs
@@ -1,4 +1,4 @@
-﻿using MyAccountingApp.Application.Services;
+﻿using MyAccountingApp.Application.DTOs;
 using MyAccountingApp.Domain.Entities;
 using MyAccountingApp.Domain.Enums;
