+++ tests/MyAccountingApp.Application.Tests/Services/CurrencyRateServiceTests.cs
@@ -1,3 +1,4 @@
+using MyAccountingApp.Application.DTOs;
 using MyAccountingApp.Application.Services;
 using MyAccountingApp.Domain.Entities;
 using MyAccountingApp.Domain.Enums;
